Initial Music Implementation
Implemented Music System Tier II:

AudioStreamInteractive Player with AudioStreamSynchronized Player child for Explore music and AudioStreamPlaylist Player for Combat music
Transitions are not fully built out at the moment due to what I assume is a bug in Godot's AudioStreamInteractive Node; The transition for next bar is not working, so the combat music currently fades out over 8 beats as the explore music fades in.  Not ideal, but it will do for now.
We will still need to work out how and where to trigger transitions between explore and combat in code